+++ react/components/EXPERIMENTAL_TableTree/Tree/TreeHeadings.tsx
@@ -19,9 +19,9 @@ const TreeHeadings: FC<TreeHeadingsProps> = ({
 }) => {
   return (
     <Row {...rowProps} height={TABLE_HEADER_HEIGHT}>
-      {columns.map((headerData: Column, headerIndex: number) => {
-        const { headerRender, title, width } = headerData
-        const content = headerRender ? headerRender({ headerData }) : title
+      {columns.map((columnData: Column, headerIndex: number) => {
+        const { headerRenderer, title, width } = columnData
+        const content = headerRenderer ? headerRenderer({ columnData }) : title
         return (
           <Row.Cell
             {...cellProps}